feat: Sidebar mobile
